requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
cloudinary==1.29.0
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
dj-database-url==1.2.0
dj-rest-auth==2.1.9
dj3-cloudinary-storage==0.0.6
Django==4.0.5
django-allauth==0.54.0
django-channels==0.7.0
django-cors-headers==3.13.0
django-countries==7.3.2
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
oauthlib==3.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyOpenSSL==22.0.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
service-identity==21.1.0
sqlparse==0.4.2
stripe==4.0.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
websockets==10.3
whitenoise==6.2.0
zope.interface==5.4.0
